Add director

+++ src/main/java/ru/yandex/practicum/filmorate/controller/film/GenreController.java
@@ -22,7 +22,7 @@ public class GenreController {
     }
 
     @GetMapping("/{id}")
-    public Genre getGenreById(@PathVariable int id) {
+    public Genre getGenreById(@PathVariable long id) {
         return genreService.getGenreById(id);
     }
 }
